requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
requests==2.32.3
huggingface_hub==0.23.4
chromadb==0.5.3
pypdf==4.2.0
langchain-text-splitters==0.2.1
python-dotenv==1.0.1
gunicorn>=21.0.0
